Added ability to manipulate windows to the UI Access plugin
You can now get a list of all windows of a specific app, just the focused window of a specific app, or all windows of all apps using a new trigger available in the Quicksilver section of the Triggers preferences.

Once you have selected a window, you can zoom, minimize, close, raise or activate it.