make demo videos more impressive specifically, add triggers which resolve the Current App, Document and Focused Window proxies, put the result in the first pane, show the interface, and then select the second pane, so you can start searching through actions right away
process objects returned from the UI Access plugin are now based on a file object representing the app, so the icon and name are correctly set, and any file or app actions can be run on the result of the Current App proxy

the Menu Bar… action has been removed because it didn't work. it was meant to let you browse menu items hierarchically, but it broke at some point. it'll be back in the rewrite.